First draft of the chat function
This is NOT the angular chat but a chat that make use of the Firebase
framework. It’s fully functional atm with the possibility to take
picture with your camera or upload pics from your computer and post in
the chat.

I will however look into the angular chat at some point because I think
this chat is kind of basic. Although you need to sign in with a Google
account to be able to use this chat.